Rename file to stop Jest from complaining
