+++ webapp/channels/src/components/advanced_text_editor/send_button/send_post_options/core_menu_options.test.tsx
@@ -5,7 +5,7 @@ import {DateTime} from 'luxon';
 import React from 'react';
 
 import useTimePostBoxIndicator from 'components/advanced_text_editor/use_post_box_indicator';
-import {WithTestMenuContext} from 'components/menu/menu_context.test';
+import {WithTestMenuContext} from 'components/menu/menu_context_test';
 
 import {renderWithContext, fireEvent, screen} from 'tests/react_testing_utils';
